+++ src/Tesseract/BitmapToPixConverter.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#if NETFULL
+
+using System;
 using System.Drawing;
 using System.Drawing.Imaging;
 
@@ -165,3 +167,5 @@ namespace Tesseract
         }
     }
 }
+
+#endif