+++ src/Tesseract/Internal/InteropDotNet/WindowsLibraryLoaderLogic.cs
@@ -3,6 +3,7 @@
 //  Distributed under the MIT License: http://opensource.org/licenses/MIT
 using System;
 using System.Runtime.InteropServices;
+using Tesseract.Internal;
 
 namespace InteropDotNet
 {
@@ -14,17 +15,17 @@ namespace InteropDotNet
 
             try
             {
-                LibraryLoaderTrace.TraceInformation("Trying to load native library \"{0}\"...", fileName);
+                Logger.TraceInformation("Trying to load native library \"{0}\"...", fileName);
                 libraryHandle = WindowsLoadLibrary(fileName);
                 if (libraryHandle != IntPtr.Zero)
-                    LibraryLoaderTrace.TraceInformation("Successfully loaded native library \"{0}\", handle = {1}.", fileName, libraryHandle);
+                    Logger.TraceInformation("Successfully loaded native library \"{0}\", handle = {1}.", fileName, libraryHandle);
                 else
-                    LibraryLoaderTrace.TraceError("Failed to load native library \"{0}\".\r\nCheck windows event log.", fileName);
+                    Logger.TraceError("Failed to load native library \"{0}\".\r\nCheck windows event log.", fileName);
             }
             catch (Exception e)
             {
                 var lastError = WindowsGetLastError();
-                LibraryLoaderTrace.TraceError("Failed to load native library \"{0}\".\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", fileName, lastError, e.ToString());
+                Logger.TraceError("Failed to load native library \"{0}\".\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", fileName, lastError, e.ToString());
             }
 
             return libraryHandle;
@@ -34,18 +35,18 @@ namespace InteropDotNet
         {
             try
             {
-                LibraryLoaderTrace.TraceInformation("Trying to free native library with handle {0} ...", libraryHandle);
+                Logger.TraceInformation("Trying to free native library with handle {0} ...", libraryHandle);
                 var isSuccess = WindowsFreeLibrary(libraryHandle);
                 if (isSuccess)
-                    LibraryLoaderTrace.TraceInformation("Successfully freed native library with handle {0}.", libraryHandle);
+                    Logger.TraceInformation("Successfully freed native library with handle {0}.", libraryHandle);
                 else
-                    LibraryLoaderTrace.TraceError("Failed to free native library with handle {0}.\r\nCheck windows event log.", libraryHandle);
+                    Logger.TraceError("Failed to free native library with handle {0}.\r\nCheck windows event log.", libraryHandle);
                 return isSuccess;
             }
             catch (Exception e)
             {
                 var lastError = WindowsGetLastError();
-                LibraryLoaderTrace.TraceError("Failed to free native library with handle {0}.\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", libraryHandle, lastError, e.ToString());
+                Logger.TraceError("Failed to free native library with handle {0}.\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", libraryHandle, lastError, e.ToString());
                 return false;
             }
         }
@@ -54,22 +55,29 @@ namespace InteropDotNet
         {
             try
             {
-                LibraryLoaderTrace.TraceInformation("Trying to load native function \"{0}\" from the library with handle {1}...",
+                Logger.TraceInformation("Trying to load native function \"{0}\" from the library with handle {1}...",
                     functionName, libraryHandle);
                 var functionHandle = WindowsGetProcAddress(libraryHandle, functionName);
                 if (functionHandle != IntPtr.Zero)
-                    LibraryLoaderTrace.TraceInformation("Successfully loaded native function \"{0}\", function handle = {1}.",
+                {
+                    Logger.TraceInformation("Successfully loaded native function \"{0}\", function handle = {1}.",
                         functionName, functionHandle);
+                }
                 else
-                    LibraryLoaderTrace.TraceError("Failed to load native function \"{0}\", function handle = {1}",
-                        functionName, functionHandle);
+                {
+                    throw new Tesseract.LoadLibraryException(String.Format(
+                        "Failed to load native function \"{0}\" from library with handle  {1}.",
+                        functionName, libraryHandle));
+
+                }
                 return functionHandle;
             }
             catch (Exception e)
             {
                 var lastError = WindowsGetLastError();
-                LibraryLoaderTrace.TraceError("Failed to free native library with handle {0}.\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", libraryHandle, lastError, e.ToString());
-                return IntPtr.Zero;
+                throw new Tesseract.LoadLibraryException(
+                    String.Format("Failed to load native function \"{0}\" from library with handle  {1}.\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", functionName, libraryHandle, lastError, e.ToString()),
+                    e);
             }
         }